adds mediatr, DI, updates structure, tests and mocks

+++ TripsApp.Api/TripsApp.Mongo/Interfaces/IFuelRateRepository.cs
@@ -4,6 +4,6 @@ namespace TripsApp.Mongo.Interfaces
 {
     public interface IFuelRateRepository : IRepository<FuelRate>
     {
-        Task<decimal> GetAsync();
+        Task<double> GetAsync();
     }
 }
